% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{calculate_cylinder_volume}
\alias{calculate_cylinder_volume}
\title{FUNCTION_TITLE}
\usage{
calculate_cylinder_volume(radius, height)
}
\arguments{
\item{radius}{PARAM_DESCRIPTION}

\item{height}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
