% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{calculate_rectangle_area}
\alias{calculate_rectangle_area}
\title{Calculate Area of a Rectangle}
\usage{
calculate_rectangle_area(length, width)
}
\arguments{
\item{length}{The length of the rectangle.}

\item{width}{The width of the rectangle.}
}
\value{
The calculated area of the rectangle.
}
\description{
This function calculates the area of a rectangle given its length and width.
}
\details{
The area of a rectangle is computed by multiplying its length by its width.
}
\examples{
\dontrun{
if(interactive()){
  calculate_rectangle_area(5, 3)  # Output: 15
}
}
}
