% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{is_even}
\alias{is_even}
\title{Check if a Number is Even}
\usage{
is_even(x)
}
\arguments{
\item{x}{The number to be checked.}
}
\value{
Logical value indicating whether the input number is even (`TRUE`) or not (`FALSE`).
}
\description{
This function checks if a given number is even.
}
\details{
This function checks whether the input number `x` is divisible by 2 without remainder.
}
\examples{
\dontrun{
if(interactive()){
  is_even(4)  # Output: TRUE
}
}
}
