% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_func.R
\name{calculate_rectangle_area}
\alias{calculate_rectangle_area}
\title{FUNCTION_TITLE}
\usage{
calculate_rectangle_area(length, width)
}
\arguments{
\item{length}{PARAM_DESCRIPTION}

\item{width}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
